src/python/trace/console/requirements.txt
opentelemetry-api==1.12.0rc2
opentelemetry-exporter-otlp-proto-grpc==1.12.0rc2
opentelemetry-sdk==1.12.0rc2
